refactor: Move core files to 'pkg' directory
- Renamed `file_organizer.go` to `pkg/file_organizer.go`
- Renamed `mime_utils.go` to `pkg/mime_utils.go`
- Renamed `tree.go` to `pkg/tree.go`
- Renamed `utils.go` to `pkg/utils.go`

This reorganization improves project structure by consolidating core files into the `pkg` directory.
