+++ pkg/file_organizer.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"encoding/json"
