+++ pkg/tree.go
@@ -1,4 +1,4 @@
-package main
+package pkg
 
 import (
 	"fmt"
